SpaceMaster/schedule-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
ariadne==0.23.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose==3.3.0
requests==2.31.0
